Update examples and restructure tests
-remove existing examples and replaced with a new 1D chain shimming example
-restructure tests into separate files in a new folder
-update optimize function to accept custom dimod solvers
-expand docstring for ``embed_bqm`` method to explain difference from existing methods in dwave-system
-promote ``optimize`` and ``make_networkx_graph`` to public methods
