+++ dwave/experimental/lattice_utils/lattice/optimize.py
@@ -20,13 +20,59 @@ import numpy as np
 import dimod
 from numpy.typing import NDArray
 
-__all__ = ['optimize', 'optimize_increasing_sa_sweeps']
+__all__ = ['optimize', 'ExponentialBackoffSimulatedAnnealingSampler']
+
+
+class ExponentialBackoffSimulatedAnnealingSampler(dimod.Sampler):
+    """SA sampler that doubles num_sweeps until energy stops improving or a cap is hit.
+
+    Starts at ``min_num_sweeps`` and doubles after each round that improves the
+    best energy, stopping when no improvement is found or ``max_num_sweeps`` is
+    exceeded.
+
+    Args:
+        max_num_sweeps: Upper bound on the number of sweeps per SA call. Once
+            ``num_sweeps`` exceeds this value, the backoff loop terminates.
+        min_num_sweeps: Initial number of sweeps for the first SA call.
+    """
+
+    properties = None
+    parameters = None
+
+    def __init__(self, max_num_sweeps=1024, min_num_sweeps=256):
+        self.sampler = SimulatedAnnealingSampler()
+        self.max_num_sweeps = max_num_sweeps
+        self.min_num_sweeps = min_num_sweeps
+        self.properties = self.sampler.properties.copy()
+        self.parameters = self.sampler.parameters.copy()
+
+    def sample(self, bqm, **parameters):
+        num_sweeps = parameters.pop("num_sweeps", self.min_num_sweeps)
+        num_reads = parameters.pop("num_reads", 256)
+
+        best_energy = np.inf
+        best_sampleset = None
+
+        while num_sweeps <= self.max_num_sweeps:
+            ss = self.sampler.sample(bqm, num_sweeps=num_sweeps, num_reads=num_reads, **parameters)
+            energy = ss.first.energy
+
+            if energy < best_energy:
+                best_energy = energy
+                best_sampleset = ss
+                num_sweeps *= 2
+            else:
+                break
+
+        best_sampleset.info["num_sweeps_exit"] = num_sweeps
+        return best_sampleset
 
 
 def optimize(
     lattice: Lattice,
     bqm: dimod.BQM,
-    sa_kwargs: dict[str, Any] | None = None,
+    sampler: dimod.Sampler | None = None,
+    sampler_kwargs: dict[str, Any] | None = None,
 ) -> tuple[float, NDArray, str]:
     """Return the best sample found by optimizing the BQM using simulated annealing.
 
@@ -44,72 +90,42 @@ def optimize(
             optimized and the resulting sample will be embedded back into the
             physical lattice.
         bqm: The binary quadratic model to optimize.
-        sa_kwargs: Optional keyword arguments to pass to the simulated annealing
-            sampler, such as ``num_reads`` and ``num_sweeps``.
+        sampler: A dimod Sampler to use for optimization of the reference energy.
+            If None, a default ExponentialBackoffSimulatedAnnealingSampler will
+            be used.
+        sampler_kwargs: Optional keyword arguments to pass to the provided
+            sampler, such as ``num_reads`` and ``num_sweeps`` in the case of a
+            SA sampler.
 
     Returns:
         A tuple containing the best energy found, the corresponding sample as a
         NumPy array, and a string indicating the optimization method used.
     """
-    if sa_kwargs is None:
-        sa_kwargs = {}
+    if sampler_kwargs is None:
+        sampler_kwargs = {}
+
+    if sampler is None:
+        sampler = ExponentialBackoffSimulatedAnnealingSampler()
+
+    reference_energy = np.inf
+    reference_sample = None
 
     # If the lattice is embedded, we should optimize the logical lattice
     if hasattr(lattice, "logical_lattice"):
         _, logical_sample, _ = optimize(
-            lattice.logical_lattice, lattice.unembed_bqm(bqm), sa_kwargs=sa_kwargs
+            lattice.logical_lattice,
+            lattice.unembed_bqm(bqm),
+            sampler=sampler,
+            sampler_kwargs=sampler_kwargs,
         )
-        embedded_sample = lattice.embed_sample(logical_sample)
-        embedded_energy = bqm.energy(embedded_sample)
+        reference_sample = lattice.embed_sample(logical_sample)
+        reference_energy = bqm.energy(reference_sample)
 
-        return optimize_increasing_sa_sweeps(bqm, embedded_energy, embedded_sample)
+    sampleset = sampler.sample(bqm, **sampler_kwargs)
+    best = sampleset.first
 
-    # If no special case, just use SA.
-    return optimize_increasing_sa_sweeps(bqm, sa_kwargs=sa_kwargs)
+    if best.energy < reference_energy:
+        sample = np.array([best.sample[v] for v in bqm.variables])
+        return best.energy, sample, type(sampler).__name__
 
-
-def optimize_increasing_sa_sweeps(
-    bqm: dimod.BQM,
-    reference_energy: float = np.inf,
-    reference_sample: NDArray | None = None,
-    sa_kwargs: dict[str, Any] | None = None,
-) -> tuple[float, NDArray, str]:
-    """Optimize a BQM with simulated annealing and increasing sweep counts.
-
-    Args:
-        bqm: The binary quadratic model to optimize.
-        reference_energy: An initial energy to compare against. If the best energy
-            found by SA is not better than this, the function will return without
-            increasing the number of sweeps.
-        reference_sample: An initial sample corresponding to the reference energy.
-        sa_kwargs: Optional keyword arguments to pass to the simulated annealing
-            sampler, such as ``num_reads`` and ``num_sweeps``. The ``num_sweeps``
-            value will be overridden by this function as it increases exponentially.
-
-    Returns:
-        A tuple containing the best energy found, the corresponding sample as a
-        NumPy array, and a string indicating the optimization method used.
-    """
-    sa = SimulatedAnnealingSampler()
-
-    if sa_kwargs is None:
-        sa_kwargs = {}
-    num_sweeps = sa_kwargs.get("num_sweeps", 256)
-    num_reads = sa_kwargs.get("num_reads", 256)
-
-    while True:
-        sample_set = sa.sample(bqm, num_reads=num_reads, num_sweeps=num_sweeps)
-        energies = sample_set.data_vectors["energy"]
-        best = np.argmin(energies)
-        best_energy = energies[best]
-
-        if best_energy < reference_energy:
-            reference_energy = best_energy
-            reference_sample = sample_set.record[best][0]
-            num_sweeps *= 2
-            if num_sweeps > 1e3:
-                break
-        else:
-            break
-
-    return reference_energy, reference_sample, "sa_exponential"
+    return reference_energy, reference_sample, type(sampler).__name__
